Add new class FilterValues and added support for it in FieldFilterBase
The new class is used to wrap a list of filter values into an object
and also support name lookups for the values.
Using a distinct class allows us to distinguish input values from
the end-user/browser and the internal values. For instance the
query might contain an identifier (e.g. 'education') which is
translated internally to an numerical ID.