Added support for custom nested filters in QuerySet
Added method `addFilter` which allows for passing a nested array structure directly to
the NestedFilter system. This requires that the filter mode is 'nested' which it is
by default.
The method may be called multiple times which makes each filter ANDed
together.
While setFilter() is easier to use the new method allows for more complex
structures with AND/OR queries.