Vastly improved filtering support for QuerySet
filter() now supports operators, they are specified on the name by
adding a colon and the operator after it. The operator must be suppored
by NestedFilter.
filter() now correctly adds attributes filters to objectFilters,
e.g. filter("folder/menu", true). This means that attribute filters no
longer needs to be defined in advanced or loaded from class.

Support for sub-queries with subQuery(), this creates a new query
object, adds it to the filter list and returns the sub-query. Makes
it easy to chain filters in sub-filters. Sub-query may also set the
filter condition (and, or, merge).
e.g.
$query
  ->subQuery(null, "or")
    ->filter("priority:>", 5)
    ->filter("priority:<", 10)
    ->parent
  ->filter("class_identifier", "folder");

Removed custom filters introduced in 1.7.0, addFilter can be used
instead.